student enrollment values and weather data start code
The student enrollment numbers are in an excel sheet and the code will
read it into a data frame.
The weather data code is not finished yet, it is currently just how to
obtain the weather for a range, the key is my key and it can only be
used 500 times a day, 10 times a minute.